Apply generic set_following to SNS.
