+++ rs/sns/governance/BUILD.bazel
@@ -27,6 +27,7 @@ DEPENDENCIES = [
     "//rs/nervous_system/common",
     "//rs/nervous_system/common/validation",
     "//rs/nervous_system/governance",
+    "//rs/nervous_system/governance/set_following",
     "//rs/nervous_system/linear_map",
     "//rs/nervous_system/lock",
     "//rs/nervous_system/proto",
